MAINT: Move `lib.format` and `ctypeslib` to dedicated submodules/private files
