+++ numpy/ctypeslib/__init__.py
@@ -1,0 +1,13 @@
+from ._ctypeslib import (
+    __all__,
+    __doc__,
+    as_array,
+    as_ctypes,
+    as_ctypes_type,
+    ctypes,
+    c_intp,
+    load_library,
+    ndpointer,
+    _concrete_ndptr,
+    _ndptr,
+)
