+++ numpy/ctypeslib/_ctypeslib.py
@@ -54,7 +54,7 @@ __all__ = ['load_library', 'ndpointer', 'c_intp', 'as_ctypes', 'as_array',
 
 import os
 import numpy as np
-from numpy._core.multiarray import _flagdict, flagsobj
+import numpy._core.multiarray as mu
 
 try:
     import ctypes
@@ -75,7 +75,9 @@ if ctypes is None:
         raise ImportError("ctypes is not available.")
     load_library = _dummy
     as_ctypes = _dummy
+    as_ctypes_type = _dummy
     as_array = _dummy
+    ndpointer = _dummy
     from numpy import intp as c_intp
     _ndptr_base = object
 else:
@@ -162,7 +164,7 @@ else:
 def _num_fromflags(flaglist):
     num = 0
     for val in flaglist:
-        num += _flagdict[val]
+        num += mu._flagdict[val]
     return num
 
 
@@ -171,7 +173,7 @@ _flagnames = ['C_CONTIGUOUS', 'F_CONTIGUOUS', 'ALIGNED', 'WRITEABLE',
 def _flags_fromnum(num):
     res = []
     for key in _flagnames:
-        value = _flagdict[key]
+        value = mu._flagdict[key]
         if (num & value):
             res.append(key)
     return res
@@ -293,7 +295,7 @@ def ndpointer(dtype=None, ndim=None, shape=None, flags=None):
         elif isinstance(flags, (int, np.integer)):
             num = flags
             flags = _flags_fromnum(num)
-        elif isinstance(flags, flagsobj):
+        elif isinstance(flags, mu.flagsobj):
             num = flags.num
             flags = _flags_fromnum(num)
         if num is None:
