+++ numpy/lib/_utils_impl.pyi
@@ -1,7 +1,10 @@
 from _typeshed import SupportsWrite
 
+from numpy._typing import DTypeLike
+
 __all__ = ["get_include", "info", "show_runtime"]
 
 def get_include() -> str: ...
 def show_runtime() -> None: ...
 def info(object: object = ..., maxwidth: int = ..., output: SupportsWrite[str] | None = ..., toplevel: str = ...) -> None: ...
+def drop_metadata(dtype: DTypeLike, /) -> DTypeLike: ...
