+++ numpy/meson.build
@@ -282,8 +282,6 @@ python_sources = [
   '_pytesttester.pyi',
   '_expired_attrs_2_0.py',
   'conftest.py',
-  'ctypeslib.py',
-  'ctypeslib.pyi',
   'exceptions.py',
   'exceptions.pyi',
   'dtypes.py',
@@ -315,6 +313,7 @@ pure_subdirs = [
   '_typing',
   '_utils',
   'compat',
+  'ctypeslib',
   'doc',
   'f2py',
   'lib',
